Remove Base template param from PublisherMixin
This is part of an effort to get rid of mixins/PublisherMixin and
mixins/ConsumerMixin.  They will be replaced with regular base classes.

Once this PR is in, I will follow up with PRs to rename
mixins/PublisherMixin and mixins/ConsumerMixin and move them to
src/automata folder.

Tested by running all the unittests