Make ServerConnectionAcceptor thread-safe
Right now it's always tied to a specific worker thread.  The issue we hit is
that there is no synchronous way of closing all FrameTransports held by the
ServerConnectionAcceptor, and waiting for all their associated events to drain
from the worker thread's EventBase.

This diff adds ServerConnectionAcceptor::stop().  This method
asynchronously calls close() on each FrameTransport, and then waits on a newly
created folly::Baton until the transport set is empty.  To make this truly safe,
we need the callbacks from the ConnectionHandler to finish executing before the
transport is removed from the set.

Using that, we can make ~RSocketServer thread-safe as well.  I've enabled
RSocketClientServer.SimpleConnect, and run it 50 times in parallel on my machine
successfully.