Add test for server binding to an invalid port
TcpConnectionAcceptor::start() should throw when it fails to bind to the port.

Note - I don't actually want to land this diff as is.  I'm looking to see if
anyone can tell me why the changes to TcpConnectionAcceptor::start() are
necessary for the test to pass in Linux/ASAN/gcc builds.