Support destroying a SetupResumeAcceptor after its EventBase is destroyed
We hit this issue when manually passing DuplexConnection objects to
RSocketServer::acceptConnection().  We generally end up with the worker threads
getting killed first, which starts destroying their EventBases and then their
folly::ThreadLocal<SetupResumeAcceptor> objects.  When we're destroying the
SetupResumeAcceptor the EventBase is gone, and accessing it will cause us to
crash.