Delete FrameBufferAllocator
This was planned on being passed into FrameSerializer so we could have the
serializer allocate frames with extra headroom for the frame length if needed.

This shouldn't be hidden by an external allocator however.  If we're adding
frame lengths, then the frame serializer should know about it and it should be
the one serializing and deserializing them.