Cleaner termination semantics for FramedReader
FramedReader is a Subscriber, so we should expect
FramedReader::on{Complete,Error}() to be called at most once, and after a call
to FramedReader::onSubscribe().

As such, we shouldn't call them from FramedReader::error(), that should have its
own logic to cancel the subscription safely and to error the inner subscriber
safely, while still being resilient to running
FramedReader::on{Complete,Error}() afterwards.  Added tests for this.

Connecting to RSocket server with `nc` and typing garbage still works properly.