Add a toString() function for FrameType
In the past we've logged these a lot (every frame received and sent), and in a
few cases we would be forced into feeding them through a std::stringstream
unnecessarily, incurring a lot of extra work.

Add a FrameType -> folly::StringPiece function in cases where the operator<<
would be too expensive, leave the operator around as it's still useful.