Build with clang-6.0 instead of clang-4.0
folly does not compile with versions of clang older than 6.0 these days.

Also set CC and CXX in the environment so that folly and gmock get built
with the intended compiler instead of being compiled with the default
system compiler (gcc 4.8.4 on Ubuntu trusty).  folly no longer compiles
with gcc 4.8 either.