getdeps: [EASY] pass yarn offline mirror location for fbsource builds
Summary:
Our CI environment cannot directly connect to the internet,
and even if it could, doing so is undesirable to fetch javascript deps.
We maintain an offline mirror of packages used by the build(s) so that
we don't have to go out to the internet.

When running in fbsource, configure the environment to use that offline
mirror.

Reviewed By: chadaustin

Differential Revision: D18061773

fbshipit-source-id: 1a5e112f23c1baaedfb3dff0c4c2a1641f6bb9a1